tests/wpt/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/moving-documents.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>When moving between documents, must refresh the original document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/multipage/semantics.html#attr-meta-http-equiv-refresh">

<div id="log"></div>

<script>
"use strict";

const sourceIFrame = document.createElement("iframe");
const destIFrame = document.createElement("iframe");
let sourceLoadCount = 0;
let destLoadCount = 0;

sourceIFrame.onload = () => {
  ++sourceLoadCount;

  if (sourceLoadCount === 2) {
    assert_equals(sourceIFrame.contentDocument.body.textContent.trim(), "foo");
    done();
  }

  maybeStartTest();
};

destIFrame.onload = () => {
  ++destLoadCount;

  if (destLoadCount === 2) {
    assert_unreached("The iframe into which the meta was moved must not refresh");
  }

  maybeStartTest();
};

function maybeStartTest() {
  if (sourceLoadCount === 1 && destLoadCount === 1) {
    const meta = sourceIFrame.contentDocument.querySelector("meta");
    destIFrame.contentDocument.body.appendChild(meta);
  }
}

sourceIFrame.src = "support/refresh.sub.html?input=" + encodeURIComponent("1; url=foo");
destIFrame.src = "support/ufoo";

document.body.appendChild(sourceIFrame);
document.body.appendChild(destIFrame);
</script>
